Fixed Specific Missiles for compatibility with Spacetime Beam
All missile pack locations are now individually selectable. The check for specific missiles has been changed to have a check for each individual pack which now checks for the corresponding item room. Due to the way that these checks now work, there is no need for `vars.missileWatchKeys` or `vars.missileMasks`, and theres lines have been removed.

Specific Missiles are sorted by area with a subtree for each(Crateria, Brinstar, etc). This seemed best since there are so many different options for Missile Packs, and it should not be necessary for other pickup types when they get added.

Room IDs for all rooms needed for these checks, as well as those that will be needed to do the same for Super Missiles, Power Bombs, and Energy/Reserve Tanks have been added.

Added a room transition split for Toilet Bowl, as requested by InsaneFirebat.

Fixed broken tooltips for `"specificSupers"` and `"specificETanks"`.

Hunted down a stray tab and replaced it with spaces.