Fixed Specific Energy Tank splits for compatibility with Spacetime Beam
Moved the `specificETanks` subtree above the `reserveTanks` option, and rewrote many of the tooltips for each energy tank to include the name of the room it is located in to facilitate easy lookup on the wiki. The order of these options has changed slightly because I rewrote the section entirely. I did this mostly because I didn't realize that all etank locations were already added, and it was easier to organize my notes and write each individual check this way. 

The `specificETanks` check has been replaced with individual checks for each etank location including roomID checks for each, and `vars.etankWatchKeys` and `vars.etankMasks` have been removed due to obsoletion.

Added an option to split on specific reserve tanks within it's own subtree. Each check was written with roomIDs for compatibility with Spacetime Beam.

Added a Rising Tide entrance split as requested by InsaneFirebat.