Adds support for various splits for Kill Samus RTA (each Ceres room and Samus hitting 0 HP) Adds support for Retroarch 1.8.0 using snes9x_libretro core.
Note about the Retroarch support:
I could have added this in the same place as the higan_sfc_libretro core, but I only tested with Retroarch 1.8.0 so I didn't want to lump it in there. I think this code needs to be re-written to support reading the WRAM address directly from the Retroarch process or from a specific module. I think this could change depending on what core you're using even if you have the same Retroarch version, but I haven't proven that.